move `getInitialNetwork` to `global-bus`
