+++ packages/global-bus/package.json
@@ -7,9 +7,12 @@
     "clear": "rm -rf build tsconfig.tsbuildinfo dist"
   },
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./util": "./src/util.ts"
   },
   "dependencies": {
+    "consts": "workspace:*",
+    "@w3ux/utils": "^2.0.10",
     "rxjs": "^7.8.2"
   },
   "devDependencies": {
